+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v3.0.0
+## 04/02/2026
+
+1. [](#new)
+    * Added 'Grav API' + Admin Pro support
+1. [](#bugfix)
+    * Fixed long codeblocks pushing content with overflow-x: auto
+    * Fixed dark mode support to theme picker modal
+
 # v2.0.2
 ## 02/03/2026